Broken: moving music to AudioService.
