+++ dev/Ultima/IO/MusicData.cs
@@ -189,25 +189,6 @@ namespace UltimaXNA.Ultima.IO
 			InternalStop ();
 		}
 
-        class UOMusic
-        {
-            public int Id;
-            public string Name;
-            public bool DoLoop;
-
-            public UOMusic()
-            {
-                Id = -1;
-                Name = "";
-                DoLoop = false;
-            }
-
-            public UOMusic(int id, string name, bool doLoop)
-            {
-                Id = id;
-                Name = name;
-                DoLoop = doLoop;
-            }
-        }
+        
 	}
 }
