+++ dev/UltimaXNA.csproj
@@ -151,6 +151,7 @@
     <Compile Include="ServiceRegistry.cs" />
     <Compile Include="Ultima\Audio\AudioService.cs" />
     <Compile Include="Ultima\Audio\SoundState.cs" />
+    <Compile Include="Ultima\Audio\UOMusic.cs" />
     <Compile Include="Ultima\Audio\UOSound.cs" />
     <Compile Include="Ultima\AUltimaModel.cs" />
     <Compile Include="Configuration\DebugSettings.cs" />
